Split coordinator/src/tributary/mod.rs into spec.rs and transaction.rs
